fix(ember): add back work item responsible filter
change export accounting file name to take it from the backend
