+++ ember/app/ui/cases/detail/download/controller.js
@@ -2,9 +2,10 @@ import Controller from "@ember/controller";
 import { inject as service } from "@ember/service";
 import { decamelize } from "@ember/string";
 import { dropTask } from "ember-concurrency-decorators";
-import { saveAs } from "file-saver";
 
 import ENV from "mysagw/config/environment";
+import downloadFile from "mysagw/utils/download-file";
+
 export default class CasesDetailDownloadController extends Controller {
   @service notification;
   @service intl;
@@ -45,23 +46,7 @@ export default class CasesDetailDownloadController extends Controller {
       headers: adapter.headers,
     };
     try {
-      const response = yield fetch(uri, init);
-
-      if (!response.ok) {
-        throw new Error(response.statusText || response.status);
-      }
-
-      const blob = yield response.blob();
-
-      // extract filename from content-disposition header e.g.
-      // inline; filename*=utf-8''2022-0948%20-%20Accus%C3%A9%20de%20r%C3%A9ception.pdf
-      const filename = decodeURIComponent(
-        response.headers
-          .get("content-disposition")
-          .match(/filename\*?=['"]?(?:UTF-\d['"]*)?([^;\r\n"']*)['"]?;?/i)[1]
-      );
-
-      saveAs(blob, filename);
+      yield downloadFile(uri, init);
     } catch (error) {
       console.error(error);
       this.notification.fromError(error);
