+++ ember/app/utils/download-file.js
@@ -1,0 +1,21 @@
+import { saveAs } from "file-saver";
+
+export default async function downloadFile(uri, init) {
+  const response = await fetch(uri, init);
+
+  if (!response.ok) {
+    throw new Error(response.statusText || response.status);
+  }
+
+  const blob = await response.blob();
+
+  // extract filename from content-disposition header e.g.
+  // inline; filename*=utf-8''2022-0948%20-%20Accus%C3%A9%20de%20r%C3%A9ception.pdf
+  const filename = decodeURIComponent(
+    response.headers
+      .get("content-disposition")
+      .match(/filename\*?=['"]?(?:UTF-\d['"]*)?([^;\r\n"']*)['"]?;?/i)[1]
+  );
+
+  saveAs(blob, filename);
+}
